3rd task

+++ delete.php
@@ -1,8 +1,14 @@
 <?php
 require "config.php";
-$id = $_GET["id"];
+if(isset($_SESSION['email']))
+{
+    $id = $_GET["id"];
 
-$sql = sprintf("DELETE FROM `users` WHERE `users`.`id` = %d" , $id );
- $result = mysqli_query($conn, $sql) or die("Query Unsuccessful");
- header("location: display.php");
+    $sql = sprintf("DELETE FROM `users` WHERE `users`.`id` = %d" , $id );
+    $result = mysqli_query($conn, $sql) or die("Query Unsuccessful");
+    header("location: display.php");
+}
+else{
+   echo  "Please login to continue"; 
+}
 ?>
